matchmaker/requirements.txt
Flask==1.1.2
redis==3.5.2
flask-redis==0.4.0
requests==2.24.0
flask-cors==3.0.9
gunicorn==20.0.4
prometheus-client==0.10.1
PyJWT==2.1.0
prometheus-flask-exporter==0.18.2